Removing html-composer references and separating out mapping layout to page data
