+++ lib/routes.js
@@ -11,7 +11,6 @@ const _ = require('lodash'),
   express = require('express'),
   path = require('path'),
   siteService = require('./services/sites'),
-  htmlComposer = require('./html-composer'),
   render = require('./render'),
   responses = require('./responses'),
   files = require('./files'),
