+++ lib/routes/components.js
@@ -7,9 +7,7 @@
 'use strict';
 
 const _ = require('lodash'),
-  // html = require('amphora-html'),
   responses = require('../responses'),
-  htmlComposer = require('../html-composer'),
   render = require('../render'),
   composer = require('../services/composer'),
   files = require('../files'),
