+++ lib/routes/pages.js
@@ -8,9 +8,9 @@
 
 const _ = require('lodash'),
   responses = require('../responses'),
-  htmlComposer = require('../html-composer'),
   render = require('../render'),
   controller = require('../services/pages'),
+  composer = require('../services/composer'),
   db = require('../services/db'),
   acceptedExtensions = {
     html: 'text/html',
@@ -24,11 +24,9 @@ const _ = require('lodash'),
  * @return {Promise}        composed data of root component and children
  */
 function getComposed(uri, locals) {
-  console.log(uri);
   return db.get(uri)
     .then(JSON.parse)
-    // require here because otherwise it's a circular dependency (via html-composer)
-    .then(pageData => require('../services/composer').composePage(pageData, locals));
+    .then(pageData => composer.composePage(pageData, locals));
 }
 
 /**
