+++ lib/services/composer.js
@@ -4,8 +4,8 @@ const _ = require('lodash'),
   bluebird = require('bluebird'),
   components = require('./components'),
   references = require('./references'),
+  mapLayoutToPageData = require('../utils/layout-to-page-data'),
   referenceProperty = '_ref';
-
 
 /**
  * Compose a component, recursively filling in all component references with
@@ -48,8 +48,7 @@ function composePage(pageData, locals) {
     pageDataNoConf = references.omitPageConfiguration(pageData);
 
   return components.get(layoutReference)
-    // require here because otherwise it's a circular dependency (via html-composer)
-    .then(layoutData => require('../html-composer').mapLayoutToPageData(pageDataNoConf, layoutData))
+    .then(layoutData => mapLayoutToPageData(pageDataNoConf, layoutData))
     .then(fullData => resolveComponentReferences(fullData, locals));
 }
